django/requirements.txt
Django==1.4
South==0.7.5
ajaxuploader==0.2.1
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.12
celery==3.0.5
distribute==0.6.24
django-browserid==0.6
django-celery==3.0.4
django-crispy-forms==1.1.3
django-dajax==0.8.4
django-dajaxice==0.2
django-debug-toolbar==0.9.4
django-extensions==0.8
django-floppyforms==0.4.7
django-registration==0.8
django-taggit==0.9.3
django-taggit-templatetags==0.4.6dev
django-templatetag-sugar==0.1
geopy==0.94.2
ipython==0.13
kombu==2.3.2
pyPdf==1.13
python-dateutil==1.5
requests==0.13.6
simplejson==2.1.6
wsgiref==0.1.2

